refactor: main to own file, functions to __init__
